refactor: mypage layout 분리
